refactor(bash): extract command-parsing/permission cluster to check.rs (dirge-4y4l stage 9b)
Second and final structural stage for the bash god-file. Moves the
command-parsing + permission-checking cluster out of bash/mod.rs into
bash/check.rs, leaving mod.rs as just the BashTool struct/impls and the
Tool impl.

- bash/check.rs (484 LOC) — check_bash_segments (raw command -> atomic
  multi-claim AccessRequest), the semantic-bash mutation helpers
  (bash_mutation_targets, mark_bash_mutations, normalize_lexical,
  fold_cd_dirs, resolve_target), and the no-semantic-bash coarse parser
  (quote_aware_split, push_segment, coarse_redirect_targets,
  coarse_mutation_paths). Both cfg paths preserved verbatim. The six
  fns reached from mod.rs/tests are pub(super); the four
  cluster-internal helpers stay private.
- bash/mod.rs (213 LOC, was 684) — BashTool + Tool impl only. call()
  references check::check_bash_segments / check::mark_bash_mutations
  fully-qualified; drops the now-unused enforce_request +
  semantic adapter imports (moved into check.rs).
- bash/tests.rs — adds `use super::check::*` for the relocated helpers.

bash.rs 2102 -> bash/{mod 213, exec 374, check 484, tests 1067}.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-y3d7
